# server/tsoa/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"RecordStatus": {
				"type": "string",
				"enum": [
					"A",
					"D"
				],
				"description": "Status ทั่วไปสำหรับ soft delete"
			},
			"Sample": {
				"description": "Sample row จาก Oracle (camelCase หลังแปลงแล้ว)",
				"properties": {
					"createdAt": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/RecordStatus"
					}
				},
				"required": [
					"id",
					"name",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SampleListResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Sample"
						},
						"type": "array"
					}
				},
				"required": [
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SampleDetailResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Sample"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SampleCreateResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"id": {
								"type": "string"
							}
						},
						"required": [
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateSampleBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateSampleBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"HealthResponse": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"ok",
							"error"
						]
					},
					"timestamp": {
						"type": "string"
					},
					"uptime": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"status",
					"timestamp",
					"uptime"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReadyResponse": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"ok",
							"error"
						]
					},
					"timestamp": {
						"type": "string"
					},
					"uptime": {
						"type": "number",
						"format": "double"
					},
					"database": {
						"type": "string",
						"enum": [
							"connected",
							"disconnected"
						]
					}
				},
				"required": [
					"status",
					"timestamp",
					"uptime",
					"database"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "fullstack-template-server",
		"version": "1.0.0",
		"contact": {}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"items": {},
											"type": "array"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get all active users",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			}
		},
		"/sample": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SampleListResponse"
								}
							}
						}
					}
				},
				"description": "ดึงรายการทั้งหมด",
				"tags": [
					"Sample"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SampleCreateResponse"
								}
							}
						}
					}
				},
				"description": "สร้างข้อมูลใหม่",
				"tags": [
					"Sample"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateSampleBody"
							}
						}
					}
				}
			}
		},
		"/sample/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SampleDetailResponse"
								}
							}
						}
					}
				},
				"description": "ดึงข้อมูลตาม ID",
				"tags": [
					"Sample"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SampleCreateResponse"
								}
							}
						}
					}
				},
				"description": "แก้ไขข้อมูล",
				"tags": [
					"Sample"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateSampleBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Remove",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"type": "number",
											"enum": [
												null
											],
											"nullable": true
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "ลบข้อมูล (soft delete)",
				"tags": [
					"Sample"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/health": {
			"get": {
				"operationId": "GetHealth",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HealthResponse"
								}
							}
						}
					}
				},
				"description": "Basic health check — always returns ok if server is running",
				"tags": [
					"Health"
				],
				"security": [],
				"parameters": []
			}
		},
		"/health/ready": {
			"get": {
				"operationId": "GetReady",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReadyResponse"
								}
							}
						}
					}
				},
				"description": "Readiness check — verifies database connectivity",
				"tags": [
					"Health"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}